Got a few boss head states in
on hit has a chance to enter shield state for a random amount of time, as phases progress this chances and time goes up. Also now has a hurt face when he takes damage. I want to implement SFX / faces for certain attacks and then music and then it should be done.